+++ QBCore_2F0666.base/resources/[qb]/qb-policejob/config.lua
@@ -1,12 +1,13 @@
 Config = {}
+-- Hard RP: UseTarget kapalı ([E] / chat RP)
 Config.UseTarget = GetConvar('UseTarget', 'false') == 'true'
-Config.MaxSpikes = 5
+Config.MaxSpikes = 3
 Config.HandCuffItem = 'handcuffs'
 Config.LicenseRank = 2
 Config.ArmoryWhitelist = {}
 Config.WhitelistedVehicles = {}
 Config.PoliceHelicopter = 'POLMAV'
-Config.FuelResource = 'LegacyFuel' -- supports any that has a GetFuel() and SetFuel() export
+Config.FuelResource = 'qb-fuel'
 
 Config.AmmoLabels = {
     AMMO_PISTOL = '9x19mm parabellum bullet',